--- src/main/java/com/ludo/study/studymatchingplatform/auth/google/service/dto/SignupRequest.java
@@ -1,5 +1,0 @@
-package com.ludo.study.studymatchingplatform.auth.google.service.dto;
-
-public class SignupRequest {
-
-}
