+++ src/main/java/com/ludo/study/studymatchingplatform/auth/service/google/GoogleLoginService.java
@@ -1,9 +1,9 @@
-package com.ludo.study.studymatchingplatform.auth.google.service;
+package com.ludo.study.studymatchingplatform.auth.service.google;
 
 import org.springframework.stereotype.Service;
 
-import com.ludo.study.studymatchingplatform.auth.google.GoogleOAuthToken;
-import com.ludo.study.studymatchingplatform.auth.google.GoogleUserInfo;
+import com.ludo.study.studymatchingplatform.auth.service.google.vo.GoogleOAuthToken;
+import com.ludo.study.studymatchingplatform.auth.service.google.vo.GoogleUserInfo;
 import com.ludo.study.studymatchingplatform.user.domain.User;
 import com.ludo.study.studymatchingplatform.user.repository.UserRepositoryImpl;
