+++ src/main/java/com/ludo/study/studymatchingplatform/auth/service/google/dto/SignupRequest.java
@@ -1,0 +1,5 @@
+package com.ludo.study.studymatchingplatform.auth.service.google.dto;
+
+public class SignupRequest {
+
+}
